Update and fixed bug

+++ app/src/main/java/com/tss/quikpawn/activities/BorrowActivity.kt
@@ -454,7 +454,7 @@ class BorrowActivity: BaseK9Activity() {
             i++
             var name = product.product_name
             var detail = product.detail
-            sumPrice += product.sale.toLong()
+            sumPrice += product.sale.toDouble()
             printerParams1 = TssPrinterParams()
             printerParams1.setAlign(PrinterParams.ALIGN.LEFT)
             printerParams1.setTextSize(20)
