+++ app/src/main/java/com/tss/quikpawn/activities/ReprintOrderActivity.kt
@@ -1499,7 +1499,7 @@ class ReprintOrderActivity : BaseK9Activity(), OrderListAdapter.OnItemClickListe
             i++
             var name = product.product_name
             var detail = product.detail
-            sumPrice += product.sale.toLong()
+            sumPrice += product.sale.toDouble()
             printerParams1 = TssPrinterParams()
             printerParams1.setAlign(PrinterParams.ALIGN.LEFT)
             printerParams1.setTextSize(20)
