+++ app/src/main/java/com/tss/quikpawn/activities/ReturnActivity.kt
@@ -326,7 +326,7 @@ class ReturnActivity : BaseK9Activity() {
             i++
             var name = product.product_name
             var detail = product.detail
-            sumPrice += product.sale.toLong()
+            sumPrice += product.sale.toDouble()
             printerParams1 = TssPrinterParams()
             printerParams1.setAlign(PrinterParams.ALIGN.LEFT)
             printerParams1.setTextSize(20)
